// src/platform/packages/shared/kbn-dom-drag-drop/src/types.d.ts
/**
 * Types of drop action
 */
export type DropType = 'field_add' | 'field_replace' | 'reorder' | 'move_compatible' | 'replace_compatible' | 'move_incompatible' | 'replace_incompatible' | 'replace_duplicate_compatible' | 'duplicate_compatible' | 'swap_compatible' | 'replace_duplicate_incompatible' | 'duplicate_incompatible' | 'swap_incompatible' | 'field_combine' | 'combine_compatible' | 'combine_incompatible';